fix(text): Fix text position with candlestick type combination
- Update conditional to calculate text position only for data specified as candlestick
- Split text test cases to separate file

Ref #2436
